Parameter moved to types/api
