+++ frontend/src/metabase-types/store/qb.ts
@@ -1,8 +1,8 @@
-import { Dataset } from "metabase-types/api/dataset";
-
 import { Card } from "metabase-types/types/Card";
 import { Field } from "metabase-types/types/Field";
-import { ParameterValueOrArray } from "metabase-types/types/Parameter";
+
+import { Dataset } from "metabase-types/api/dataset";
+import { ParameterValueOrArray } from "metabase-types/api/parameter";
 
 export type QueryBuilderMode = "view" | "notebook" | "dataset";
 export type DatasetEditorTab = "query" | "metadata";
